+++ app/src/main/java/com/money/manager/ex/domainmodel/Stock.java
@@ -87,8 +87,8 @@ public class Stock
 
     // properties
 
-    public Integer getId() {
-        return getInt(StockFields.STOCKID);
+    public Long getId() {
+        return getLong(StockFields.STOCKID);
     }
 
     public Money getCommission() {
@@ -108,12 +108,12 @@ public class Stock
         contentValues.put(StockFields.CURRENTPRICE, currentPrice.toString());
     }
 
-    public int getHeldAt() {
-        return getInt(StockFields.HELDAT);
+    public long getHeldAt() {
+        return getLong(StockFields.HELDAT);
     }
 
-    public void setHeldAt(int value) {
-        setInt(StockFields.HELDAT, value);
+    public void setHeldAt(long value) {
+        setLong(StockFields.HELDAT, value);
     }
 
     public String getNotes() {
